Refactor GitHubIntegrationContextType
